line_art_generator/api/requirements.txt
werkzeug
Pillow
